+++ Pinta.Core/SettingNames.cs
@@ -8,6 +8,8 @@ internal static class SettingNames
 
 	internal const string WEBP_QUALITY = "webp-quality";
 
+	internal const string AVIF_QUALITY = "avif-quality";
+
 	internal const string SELECTION_COMBINE_MODE = "selection-combine-mode";
 
 	internal const string SHOW_CANVAS_GRID = "show-canvas-grid";