+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Thanks to the following contributors who worked on this release:
 
 ### Added
 - The splatter brush now allows the minimum and maximum splatter size to be configured separately from the brush width
+- Impasto can now save images as AVIF (#2185)
 - Added a new "Slash" brush type (#1974)
 - The status bar color palette now supports Ctrl+clicking to edit a color, in addition to middle clicking (#1436)
 - The Resize Image and Resize Canvas dialogs now remember their settings (#1869, #1972)